Unit Unavaliable Animation
Added method to set unit to unavailable
Added Guid to Unit for easier Find(). This is a better solution in the long run as multiple units could have the same values in the same map
Added bool for if Unit is Active (ie. can it be used)
Removed one of the Units() constructors as it was un-necessary as the other constructor does the same thing